Renamed esm scripts to end with mjs
