+++ scripts/camera/base-camera.mjs
